remove FaqListAdapter and put this adapter as a inner class
